+++ switzerland/src/main/java/org/eqasim/switzerland/scenario/RunAdaptConfig.java
@@ -1,18 +1,21 @@
 package org.eqasim.switzerland.scenario;
 
-import org.eqasim.core.components.config.ConfigAdapter;
 import org.eqasim.core.components.config.EqasimConfigGroup;
 import org.eqasim.switzerland.SwitzerlandConfigurator;
 import org.eqasim.switzerland.mode_choice.SwissModeChoiceModule;
 import org.matsim.api.core.v01.TransportMode;
 import org.matsim.contribs.discrete_mode_choice.modules.config.DiscreteModeChoiceConfigGroup;
+import org.matsim.core.config.*;
 import org.matsim.core.config.CommandLine.ConfigurationException;
-import org.matsim.core.config.Config;
+import org.matsim.core.config.groups.PlanCalcScoreConfigGroup;
+
+
 
 public class RunAdaptConfig {
+
 	static public void main(String[] args) throws ConfigurationException {
 		SwitzerlandConfigurator configurator = new SwitzerlandConfigurator();
-		ConfigAdapter.run(args, configurator.getConfigGroups(), RunAdaptConfig::adaptConfiguration);
+		SwissConfigAdapter.run(args, configurator.getConfigGroups(), RunAdaptConfig::adaptConfiguration);
 	}
 
 	static public void adaptConfiguration(Config config) {
@@ -28,5 +31,23 @@ public class RunAdaptConfig {
 				.get(DiscreteModeChoiceConfigGroup.GROUP_NAME);
 
 		dmcConfig.setModeAvailability(SwissModeChoiceModule.MODE_AVAILABILITY_NAME);
+
+		// adapting Scoring config with custom activities
+		if (SwissConfigAdapter.hasCustomActivities) {
+			PlanCalcScoreConfigGroup scoringConfig = config.planCalcScore();
+
+			for (String activityType : SwissConfigAdapter.activityTypes) {
+				PlanCalcScoreConfigGroup.ActivityParams activityParams = scoringConfig.getActivityParams(activityType);
+
+				if (activityParams == null) {
+					activityParams = new PlanCalcScoreConfigGroup.ActivityParams(activityType);
+					config.planCalcScore().addActivityParams(activityParams);
+				}
+
+				activityParams.setScoringThisActivityAtAll(false);
+			}
+		}
+
 	}
+
 }